requirements.txt
aiogram==3.4.1
apscheduler==3.10.4
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
pytz==2023.3
openai==1.6.1
dateparser==1.2.0
aiofiles==23.2.1